app/requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework_simplejwt==4.6.0
weasyprint==50
psycopg2-binary==2.8.3
Pillow>=5.3.0,<5.4.0
django-cors-headers>=3.2.0,<3.3.0
djangorestframework-jwt==1.11.0
requests==2.22.0
sendgrid==6.1.1
django-environ==0.4.5
django-crispy-forms==1.8.1
gunicorn==20.0.4
schedule==0.6.0
whitenoise==5.0.1
drf-api-logger==1.0.6
django-simple-history==3.0.0
bahttext==1.0.2
python-dateutil==2.8.2